Move kolibri wrapper to a Python package
