+++ kolibri-flatpak-utils/content_extensions.py
@@ -1,28 +1,14 @@
-#!/usr/bin/python3
-
-# Kolibri flatpak wrapper script.
-# Discovers content extension directories, sets KOLIBRI_CONTENT_FALLBACK_DIRS
-# accordingly, and runs scanforcontent when content extensions have changed.
-
-import logging
-logger = logging.getLogger(__name__)
-
-import datetime
 import json
 import operator
 import os
 import re
-import subprocess
-import sys
 
-from collections import namedtuple
 from configparser import ConfigParser
 
-from kolibri_gnome.globals import init_logging, KOLIBRI_HOME
+from kolibri_gnome.globals import KOLIBRI_HOME
 
 
-KOLIBRI = os.environ.get('X_KOLIBRI', 'kolibri')
-CONTENT_EXTENSIONS_DIR = os.environ.get('X_KOLIBRI_CONTENT_EXTENSIONS_DIR', '')
+CONTENT_EXTENSIONS_DIR = '/app/share/kolibri-content'
 
 CONTENT_EXTENSION_RE = r'^org.learningequality.Kolibri.Content.(?P<name>\w+)$'
 
@@ -221,91 +207,4 @@ class ContentExtensionsList(object):
 
     def __filter_extensions(self, refs):
         return set(map(self.__get_extension, refs))
-
-
-class Application(object):
-    def __init__(self):
-        self.__cached_extensions = ContentExtensionsList.from_cache()
-        self.__active_extensions = ContentExtensionsList.from_flatpak_info()
-
-    def run(self):
-        process_success = all([
-            self.__process_removed_extensions(),
-            self.__process_added_extensions(),
-            self.__process_updated_extensions()
-        ])
-
-        if process_success:
-            self.__active_extensions.write_to_cache()
-
-    def __process_removed_extensions(self):
-        # For each removed channel, run scanforcontent
-        # TODO: Is there a less expensive way of doing this than scanforcontent?
-        channel_ids = set()
-        for extension in ContentExtensionsList.removed(self.__cached_extensions, self.__active_extensions):
-            logging.info("Removed extension: %s", extension.ref)
-            channel_ids.update(
-                map(operator.attrgetter('channel_id'), extension.channels)
-            )
-        return self.__kolibri_scan_content(channel_ids, ['--channel-import-mode=none'])
-
-    def __process_added_extensions(self):
-        # For each added channel, run scanforcontent
-        # TODO: Instead of scanforcontent, use importcontent with --node_ids and --exclude_node_ids
-        channel_ids = set()
-        for extension in ContentExtensionsList.added(self.__cached_extensions, self.__active_extensions):
-            logging.info("Added extension: %s", extension.ref)
-            channel_ids.update(
-                map(operator.attrgetter('channel_id'), extension.channels)
-            )
-        return self.__kolibri_scan_content(channel_ids)
-
-    def __process_updated_extensions(self):
-        # For each updated channel, run scanforcontent
-        # TODO: Instead of scanforcontent, use importcontent with --node_ids and --exclude_node_ids
-        channel_ids = set()
-        for extension in ContentExtensionsList.updated(self.__cached_extensions, self.__active_extensions):
-            logging.info("Updated extension: %s", extension.ref)
-            channel_ids.update(
-                map(operator.attrgetter('channel_id'), extension.channels)
-            )
-        return self.__kolibri_scan_content(channel_ids)
-
-    def __kolibri_scan_content(self, channel_ids, args=[]):
-        if len(channel_ids) == 0:
-            return True
-
-        logger.info("scanforcontent starting for channels %s: %s", channel_ids, datetime.datetime.today())
-
-        try:
-            self.__run_kolibri(['manage', 'scanforcontent', '--channels={}'.format(','.join(channel_ids)), *args])
-        except CalledProcessException as error:
-            logger.warning("scanforcontent failed: %s", datetime.datetime.today())
-            return False
-        else:
-            logger.info("scanforcontent completed: %s", datetime.datetime.today())
-            return True
-
-    def __run_kolibri(self, args=None):
-        if args is None:
-            args = sys.argv[1:]
-        result = subprocess.run([KOLIBRI, *args], env=self.__kolibri_env, check=True)
-        return result.returncode
-
-    @property
-    def __kolibri_env(self):
-        kolibri_env = os.environ.copy()
-        kolibri_env['KOLIBRI_CONTENT_FALLBACK_DIRS'] = ';'.join(self.__active_extensions.content_fallback_dirs)
-        return kolibri_env
-
-
-def main():
-    init_logging('kolibri-flatpak-wrapper.txt')
-    application = Application()
-    return application.run()
-
-
-if __name__ == "__main__":
-    result = main()
-    sys.exit(result)
